terminado a falta de un cambio visual

+++ Controller/SubController.java
@@ -5,6 +5,7 @@ import Model.Repos.RepoProject;
 import Model.Repos.RepoUser;
 import Model.Session;
 import Model.Task;
+import Model.User;
 import Utils.IO;
 import View.MainView;
 import View.SubView;
@@ -21,6 +22,7 @@ public class SubController {
     RepoUser repoUser = RepoUser.getInstance();
     RepoProject repoProject = RepoProject.getInstance();
     TabsView tabsView = new TabsView();
+
 
     /**
      * Menu de nuestra primera opción
@@ -91,9 +93,18 @@ public class SubController {
                     break;
                 case 4: // Borrar proyecto
                     String projectName = IO.readString("Introduce el nombre del proyecto que quieres borrar: ");
-                    Project deleteProject = repoProjectController.getProject(projectName);
-                    repoProjectController.deleteProject(deleteProject);
-                    saveAll();
+                    Project projectToDelete = repoProjectController.getProject(projectName);
+                    if (projectToDelete != null) {
+                        if (projectToDelete.isCreator(Session.getInstance().getLoggedInUser())) {
+                            // El usuario es el creador del proyecto
+                            repoProjectController.deleteProject(projectToDelete);
+                            saveAll();
+                        } else {
+                            System.out.println("Solo el creador del proyecto puede eliminar el proyecto.");
+                        }
+                    } else {
+                        System.out.println("El proyecto no existe.");
+                    }
                     break;
             }
         } while (option != 5);
@@ -167,39 +178,58 @@ public class SubController {
                 case 1: // Cambiar nombre
                     String projectName = IO.readString("Introduce el nombre del proyecto que quieres modificar: ");
                     Project project = repoProjectController.getProject(projectName);
-                    if (repoProjectController.projectExists(project.getName())) {
-                        String newProjectName = IO.readString("Introduce el nuevo nombre del proyecto: ");
-                        repoProjectController.updateName(project.getName(), newProjectName);
-                        MainView.showMessage("El nombre del proyecto ha sido modificado exitosamente");
-                        saveAll();
+                    if (project != null) {
+                        if (project.isCreator(Session.getInstance().getLoggedInUser())) {
+                            String newProjectName = IO.readString("Introduce el nuevo nombre del proyecto: ");
+                            repoProjectController.updateName(project.getName(), newProjectName);
+                            MainView.showMessage("El nombre del proyecto ha sido modificado exitosamente");
+                            saveAll();
+                        } else {
+                            MainView.showMessage("Solo el creador del proyecto puede cambiar el nombre.");
+                        }
                     } else {
                         MainView.showMessage("El proyecto no existe");
                     }
                     break;
                 case 2: // Añadir colaborador
-                    String projectToModify = IO.readString("Introduce el nombre del proyecto al que añadir el colaborador: ");
-                    if (projectToModify.equals(repoProjectController.getProject(projectToModify).getName())) {
-                        String collaboratorToAdd = IO.readString("Introduce el nombre de usuario del colaborador: ");
-                        if (userController.userExists(collaboratorToAdd)) {
-                            repoProjectController.addCollaborator(repoProjectController.getProject(projectToModify), userController.getUser(collaboratorToAdd));
-                            MainView.showMessage("El colaborador "+ collaboratorToAdd +" ha sido añadido exitosamente");
-                            saveAll();
+                    String projectNameToAddCollaborator = IO.readString("Introduce el nombre del proyecto al que añadir el colaborador: ");
+                    Project projectToAddCollaborator = repoProjectController.getProject(projectNameToAddCollaborator);
+                    if (projectToAddCollaborator != null) {
+                        if (projectToAddCollaborator.isCreator(Session.getInstance().getLoggedInUser())) {
+                            userController.showAllUsers();
+                            String collaboratorToAdd = IO.readString("Introduce el nombre de usuario del colaborador: ");
+                            if (userController.userExists(collaboratorToAdd)) {
+                                repoProjectController.addCollaborator(projectToAddCollaborator, userController.getUser(collaboratorToAdd));
+                                saveAll();
+                            } else {
+                                MainView.showMessage("El usuario no existe");
+                            }
                         } else {
-                            MainView.showMessage("El colaborador no existe");
+                            MainView.showMessage("Solo el creador del proyecto puede añadir un colaborador.");
                         }
+                    } else {
+                        MainView.showMessage("El proyecto no existe");
                     }
                     break;
                 case 3: // Borrar colaborador
                     String projectModify = IO.readString("Introduce el nombre del proyecto al que añadir el colaborador: ");
-                    if (projectModify.equals(repoProjectController.getProject(projectModify).getName())) {
-                        String collaboratorToRemove = IO.readString("Introduce el nombre de usuario del colaborador: ");
-                        if (userController.userExists(collaboratorToRemove)) {
-                            repoProjectController.removeCollaborator(repoProjectController.getProject(projectModify), userController.getUser(collaboratorToRemove));
-                            MainView.showMessage("El colaborador ha sido eliminado exitosamente");
-                            saveAll();
+                    Project projectToRemoveCollaborator = repoProjectController.getProject(projectModify);
+                    if (projectToRemoveCollaborator != null) {
+                        if (projectToRemoveCollaborator.isCreator(Session.getInstance().getLoggedInUser())) {
+                            repoProjectController.showAllCollaborators(projectToRemoveCollaborator);
+                            String collaboratorToRemove = IO.readString("Introduce el nombre del colaborador que quieres eliminar: ");
+                            if (userController.userExists(collaboratorToRemove)) {
+                                repoProjectController.removeCollaborator(projectToRemoveCollaborator, userController.getUser(collaboratorToRemove));
+                                MainView.showMessage("El colaborador ha sido eliminado exitosamente");
+                                saveAll();
+                            } else {
+                                MainView.showMessage("El colaborador no existe");
+                            }
                         } else {
-                            MainView.showMessage("El colaborador no existe");
+                            MainView.showMessage("Solo el creador del proyecto puede eliminar un colaborador.");
                         }
+                    } else {
+                        MainView.showMessage("El proyecto no existe");
                     }
                     break;
                 case 4: // Modificar tareas
@@ -219,8 +249,13 @@ public class SubController {
                 case 1: // Añadir tarea
                     String projectTask = IO.readString("Introduce el nombre del proyecto al que quieres añadir una tarea: ");
                     if (repoProjectController.projectExists(projectTask)) {
-                        projectController.createTask(repoProjectController.getProject(projectTask));
-                        saveAll();
+                        Project project = repoProjectController.getProject(projectTask);
+                        if (project.isCreator(Session.getInstance().getLoggedInUser())) {
+                            projectController.createTask(project);
+                            saveAll();
+                        } else {
+                            MainView.showMessage("Solo el creador del proyecto puede añadir una tarea.");
+                        }
                     } else {
                         MainView.showMessage("El proyecto no existe");
                     }
@@ -237,8 +272,8 @@ public class SubController {
                             case 1:
                                 String projectTaskShow = IO.readString("Introduce el nombre del proyecto al que quieres mostrar una tarea: ");
                                 Project project = repoProjectController.getProject(projectTaskShow);
-                                    String taskToShow = IO.readString("Introduce el nombre de la tarea que quieres mostrar: ");
-                                    projectController.showTask(project, taskToShow);
+                                String taskToShow = IO.readString("Introduce el nombre de la tarea que quieres mostrar: ");
+                                projectController.showTask(project, taskToShow);
 
                                 break;
                             case 2:
@@ -255,10 +290,13 @@ public class SubController {
                     if (repoProjectController.projectExists(projectUName.getName())){
                         String taskToRename = IO.readString("Introduce el nombre de la tarea que quieres cambiar el nombre: ");
                         String newTaskName = IO.readString("Introduce el nuevo nombre de la tarea: ");
-                        projectController.updateTaskName(projectUName, taskToRename, newTaskName);
-                        MainView.showMessage("El nombre de la tarea ha sido cambiado exitosamente");
-                        saveAll();
-                    }else {
+                        if (projectUName.isCreator(Session.getInstance().getLoggedInUser())) {
+                            projectController.updateTaskName(projectUName, taskToRename, newTaskName);
+                            saveAll();
+                        } else {
+                            MainView.showMessage("Solo el creador del proyecto puede cambiar el nombre de una tarea");
+                        }
+                    } else {
                         MainView.showMessage("El proyecto no existe");
                     }
                     break;
@@ -268,9 +306,13 @@ public class SubController {
                     Project projectTDelete = repoProjectController.getProject(projectTaskDelete);
                     if (repoProjectController.projectExists(projectTDelete.getName())) {
                         String taskToDelete = IO.readString("Introduce el nombre de la tarea que quieres borrar: ");
-                        projectController.deleteTask(projectTDelete, taskToDelete);
-                        saveAll();
-                        MainView.showMessage("La tarea ha sido eliminada exitosamente");
+                        if (projectTDelete.isCreator(Session.getInstance().getLoggedInUser())) {
+                            projectController.deleteTask(projectTDelete, taskToDelete);
+                            saveAll();
+                            MainView.showMessage("La tarea ha sido eliminada exitosamente");
+                        } else {
+                            MainView.showMessage("Solo el creador del proyecto puede eliminar una tarea");
+                        }
                     } else {
                         MainView.showMessage("El proyecto no existe");
                     }
@@ -278,36 +320,85 @@ public class SubController {
                 case 5: // Cambiar el estado
                     String projectTaskStatus = IO.readString("Introduce el nombre del proyecto al que quieres cambiar el estado de una tarea: ");
                     Project project = repoProjectController.getProject(projectTaskStatus);
-                    String taskToChangeStatus = IO.readString("Introduce el nombre de la tarea que quieres cambiar el estado: ");
-                    projectController.changeTaskStatus(project, taskToChangeStatus);
-                    saveAll();
+                    if (project != null) {
+                        String taskToChangeStatus = IO.readString("Introduce el nombre de la tarea que quieres cambiar el estado: ");
+                        Task selectedTask = null;
+                        for (Task task : project.getTasks()) {
+                            if (task.getName().equals(taskToChangeStatus)) {
+                                selectedTask = task;
+                                break;
+                            }
+                        }
+                        if (selectedTask != null && (project.isCreator(Session.getInstance().getLoggedInUser()) ||
+                                selectedTask.getAssignedUser().equals(Session.getInstance().getLoggedInUser().getUsername()))) {
+                            projectController.changeTaskStatus(project, taskToChangeStatus);
+                            saveAll();
+                        } else if (selectedTask == null) {
+                            MainView.showMessage("La tarea no existe en el proyecto");
+                        } else {
+                            MainView.showMessage("Solo el creador del proyecto o el usuario asignado a la tarea pueden cambiar el estado");
+                        }
+                    } else {
+                        MainView.showMessage("El proyecto no existe");
+                    }
                     break;
                 case 6: // Cambiar usuario asignado
                     String projectTaskChangeUser = IO.readString("Introduce el nombre del proyecto al que quieres cambiar el usuario asignado de la tarea: ");
                     Project projectUsers = repoProjectController.getProject(projectTaskChangeUser);
-                    String taskToChangeUser = IO.readString("Introduce el nombre de la tarea que quieres cambiar el usuario asignado: ");
-                    Task task = projectController.getTask(projectUsers, taskToChangeUser);
-                    repoProjectController.showAllCollaborators(projectUsers);
-                    String changeUserAssigned = IO.readString("Introduce el nombre del usuario asignado nuevo: ");
-                    projectController.updateAssignedUser(projectUsers,task, changeUserAssigned);
-                    saveAll();
+
+                    if (projectUsers != null) {
+                        String taskToChangeUser = IO.readString("Introduce el nombre de la tarea que quieres cambiar el usuario asignado: ");
+                        Task task = projectController.getTask(projectUsers, taskToChangeUser);
+
+                        if (task != null) {
+                            // Verificar si el usuario conectado es el creador del proyecto
+                            if (projectUsers.isCreator(Session.getInstance().getLoggedInUser())) {
+                                repoProjectController.showAllCollaborators(projectUsers);
+                                String changeUserAssigned = IO.readString("Introduce el nombre del usuario asignado nuevo: ");
+                                // Verificar si el usuario nuevo asignado es un colaborador del proyecto
+                                if (repoProjectController.collaboratorExists(projectUsers, changeUserAssigned)) {
+                                    projectController.updateAssignedUser(projectUsers, task, changeUserAssigned);
+                                    saveAll();
+                                } else {
+                                    MainView.showMessage("El usuario asignado debe ser un colaborador del proyecto.");
+                                }
+                            } else {
+                                MainView.showMessage("Solo el creador del proyecto puede cambiar el usuario asignado de la tarea.");
+                            }
+                        } else {
+                            MainView.showMessage("Esta tarea no existe.");
+                        }
+                    } else {
+                        MainView.showMessage("Este proyecto no existe.");
+                    }
+
                     break;
                 case 7: // Crear comentario
                     String projectTaskComment = IO.readString("Introduce el nombre del proyecto al que quieres añadir un comentario: ");
                     Project projectComment = repoProjectController.getProject(projectTaskComment);
+
                     if (projectComment != null) {
                         String taskToComment = IO.readString("Introduce el nombre de la tarea que quieres comentar: ");
                         Task taskComment = projectController.getTask(projectComment, taskToComment);
+
                         if (taskComment != null) {
-                            String comment = IO.readString("Introduce el comentario a añadir: ");
-                            projectController.createComment(projectComment, taskComment, comment);
-                            saveAll();
+                            User loggedInUser = Session.getInstance().getLoggedInUser();
+                            // Verificar si el usuario conectado es el creador del proyecto o está asignado a la tarea
+                            if (projectComment.isCreator(loggedInUser) ||
+                                    taskComment.getAssignedUser().equals(Session.getInstance().getLoggedInUser().getUsername())) {
+                                String comment = IO.readString("Introduce el comentario a añadir: ");
+                                projectController.createComment(projectComment, taskComment, comment);
+                                saveAll();
+                            } else {
+                                MainView.showMessage("Solo el creador del proyecto o los usuarios asignados a la tarea pueden añadir comentarios.");
+                            }
                         } else {
                             MainView.showMessage("La tarea no existe en el proyecto.");
                         }
                     } else {
                         MainView.showMessage("El proyecto no existe.");
                     }
+                    break;
             }
         } while (option != 8);
     }
